feat: split passthrough flashers into tabs
